refactor: simplify component filenames
