+++ src/app/(home)/_components/about.tsx
@@ -1,12 +1,12 @@
 import Link from 'next/link'
 import { Prose } from '@/components/content/mdx/prose'
 import { Icons } from '@/components/icons/icons'
-import { SectionHeader } from '@/components/layout/sections/section-header'
+import { SectionHeader } from '@/components/layout/sections/header'
 import {
   SplitSection,
   SplitSectionContent,
   SplitSectionSidebar,
-} from '@/components/layout/sections/split-section'
+} from '@/components/layout/sections/split'
 import { buttonVariants } from '@/components/ui/button'
 import { ViewAnimation } from '@/components/view-animation'
 import { cn } from '@/lib/utils'
