+++ src/app/(home)/_components/hero.tsx
@@ -3,7 +3,7 @@ import * as motion from 'motion/react-client'
 import Image from 'next/image'
 import Link from 'next/link'
 import { Icons } from '@/components/icons/icons'
-import { Section } from '@/components/layout/sections/section'
+import { Section } from '@/components/layout/sections'
 import { buttonVariants } from '@/components/ui/button'
 import { ViewAnimation } from '@/components/view-animation'
 import { linkItems } from '@/constants/navigation'
