+++ src/app/(home)/_components/posts.tsx
@@ -1,6 +1,6 @@
 import type { HTMLAttributes } from 'react'
 import { PostCard } from '@/components/features/blog/post-card'
-import { Section } from '@/components/layout/sections/section'
+import { Section } from '@/components/layout/sections'
 import { SearchRedirectInput } from '@/components/search-redirect-input'
 import { ViewAnimation } from '@/components/view-animation'
 import type { BlogPage } from '@/lib/source'
