+++ src/app/(home)/_components/testimonials.tsx
@@ -2,8 +2,8 @@
 
 import { useEffect, useState } from 'react'
 import { Icons } from '@/components/icons/icons'
-import { Section } from '@/components/layout/sections/section'
-import { SectionHeader } from '@/components/layout/sections/section-header'
+import { Section } from '@/components/layout/sections'
+import { SectionHeader } from '@/components/layout/sections/header'
 import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar'
 import {
   Carousel,
