+++ src/app/(home)/_components/work.tsx
@@ -1,8 +1,8 @@
 import Link from 'next/link'
 import { WorkCard } from '@/components/features/work/work-card'
 import { Icons } from '@/components/icons/icons'
-import { Section } from '@/components/layout/sections/section'
-import { SectionHeader } from '@/components/layout/sections/section-header'
+import { Section } from '@/components/layout/sections'
+import { SectionHeader } from '@/components/layout/sections/header'
 import { buttonVariants } from '@/components/ui/button'
 import { ViewAnimation } from '@/components/view-animation'
 import type { WorkPage } from '@/lib/source'
