+++ src/app/(home)/about/_components/detailed.tsx
@@ -5,7 +5,7 @@ import {
   SplitSectionContent,
   SplitSectionHeader,
   SplitSectionSidebar,
-} from '@/components/layout/sections/split-section'
+} from '@/components/layout/sections/split'
 import { buttonVariants } from '@/components/ui/button'
 import { ViewAnimation } from '@/components/view-animation'
 import { cn } from '@/lib/utils'
