+++ src/app/(home)/blog/[slug]/_components/header.tsx
@@ -2,7 +2,7 @@
 import * as motion from 'motion/react-client'
 import { BlurImage } from '@/components/blur-image'
 import { TagCard } from '@/components/features/blog/tag-card'
-import { Section } from '@/components/layout/sections/section'
+import { Section } from '@/components/layout/sections'
 import { ViewAnimation } from '@/components/view-animation'
 import type { BlogPage as MDXPage } from '@/lib/source'
