+++ src/app/(home)/blog/_components/newsletter-section.tsx
@@ -4,7 +4,7 @@ import {
   SplitSectionContent,
   SplitSectionHeader,
   SplitSectionSidebar,
-} from '@/components/layout/sections/split-section'
+} from '@/components/layout/sections/split'
 import { ViewAnimation } from '@/components/view-animation'
 
 export function NewsletterSection() {
