+++ src/app/(home)/blog/page.tsx
@@ -2,8 +2,8 @@ import type { Metadata, ResolvingMetadata } from 'next'
 import { notFound, redirect } from 'next/navigation'
 import Posts from '@/app/(home)/_components/posts'
 import { CollectionPageJsonLd } from '@/components/json-ld'
-import { Section } from '@/components/layout/sections/section'
-import { SectionBody } from '@/components/layout/sections/section-body'
+import { Section } from '@/components/layout/sections'
+import { SectionBody } from '@/components/layout/sections/body'
 import { Wrapper } from '@/components/layout/wrapper'
 import { NumberedPagination } from '@/components/numbered-pagination'
 import { postsPerPage } from '@/constants/config'
