+++ src/app/(home)/blog/tags/[...slug]/page.tsx
@@ -3,9 +3,9 @@ import { notFound, redirect } from 'next/navigation'
 import { PostCard } from '@/components/features/blog/post-card'
 import { Icons } from '@/components/icons/icons'
 import { TagJsonLd } from '@/components/json-ld'
+import { Section } from '@/components/layout/sections'
+import { SectionBody } from '@/components/layout/sections/body'
 import { HeroSection } from '@/components/layout/sections/hero'
-import { Section } from '@/components/layout/sections/section'
-import { SectionBody } from '@/components/layout/sections/section-body'
 import { NumberedPagination } from '@/components/numbered-pagination'
 import { ViewAnimation } from '@/components/view-animation'
 import { postsPerPage } from '@/constants/config'
