+++ src/app/(home)/guestbook/page.tsx
@@ -1,13 +1,13 @@
 import type { Metadata } from 'next'
 import { Suspense } from 'react'
 import { WebPageJsonLd } from '@/components/json-ld'
-import { Section } from '@/components/layout/sections/section'
+import { Section } from '@/components/layout/sections'
 import {
   SplitSection,
   SplitSectionContent,
   SplitSectionHeader,
   SplitSectionSidebar,
-} from '@/components/layout/sections/split-section'
+} from '@/components/layout/sections/split'
 import { Wrapper } from '@/components/layout/wrapper'
 import { ViewAnimation } from '@/components/view-animation'
 import { createMetadata } from '@/lib/metadata'
