+++ src/app/(home)/uses/page.tsx
@@ -2,8 +2,8 @@ import type { Metadata } from 'next'
 import type { ReactNode } from 'react'
 import { Fragment } from 'react'
 import { AboutPageJsonLd } from '@/components/json-ld'
-import { Section } from '@/components/layout/sections/section'
-import Separator from '@/components/layout/separator'
+import { Section } from '@/components/layout/sections'
+import Separator from '@/components/layout/sections/separator'
 import { Wrapper } from '@/components/layout/wrapper'
 import { ViewAnimation } from '@/components/view-animation'
 import { hardware, software } from '@/constants/portfolio/uses'
