+++ src/app/(home)/work/[slug]/page.tsx
@@ -9,8 +9,8 @@ import { PreviewButton } from '@/components/content/mdx/preview-button'
 import { TOCMinimap, TOCPopover } from '@/components/features/toc'
 import { Icons } from '@/components/icons/icons'
 import { WorkJsonLd } from '@/components/json-ld'
-import { Section } from '@/components/layout/sections/section'
-import { SectionBody } from '@/components/layout/sections/section-body'
+import { Section } from '@/components/layout/sections'
+import { SectionBody } from '@/components/layout/sections/body'
 import { ViewAnimation } from '@/components/view-animation'
 import { description as homeDescription } from '@/constants/site'
 import { createMetadata, getWorkPageImage } from '@/lib/metadata'
