+++ src/app/(home)/work/page.tsx
@@ -2,7 +2,7 @@ import type { Metadata, ResolvingMetadata } from 'next'
 import { notFound, redirect } from 'next/navigation'
 import { WorkCard } from '@/components/features/work/work-card'
 import { CollectionPageJsonLd } from '@/components/json-ld'
-import { Section } from '@/components/layout/sections/section'
+import { Section } from '@/components/layout/sections'
 import { Wrapper } from '@/components/layout/wrapper'
 import { NumberedPagination } from '@/components/numbered-pagination'
 import { SearchRedirectInput } from '@/components/search-redirect-input'
