+++ src/app/provider.tsx
@@ -7,9 +7,9 @@ import type { ReactNode } from 'react'
 import Analytics from '@/components/analytics'
 import { AISearch } from '@/components/features/assistant/chat'
 import { AISearchTrigger } from '@/components/features/assistant/search-trigger'
-import { MobileNav } from '@/components/layout/header/mobile/mobile-nav'
-import { SmoothCursor } from '@/components/providers/smooth-cursor'
-import { TailwindIndicator } from '@/components/providers/tailwind-indicator'
+import { MobileNav } from '@/components/layout/header/mobile'
+import { SmoothCursor } from '@/components/smooth-cursor'
+import { TailwindIndicator } from '@/components/tailwind-indicator'
 import { Toaster } from '@/components/ui/sonner'
 import { TooltipProvider } from '@/components/ui/tooltip'
