+++ src/components/layout/sections/hero.tsx
@@ -2,7 +2,7 @@ import { cva, type VariantProps } from 'class-variance-authority'
 import { Children, type ReactNode } from 'react'
 import { ViewAnimation } from '@/components/view-animation'
 import { cn } from '@/lib/utils'
-import { Section } from './section'
+import { Section } from './index'
 
 const heroVariants = cva('flex flex-col', {
   defaultVariants: {
