+++ src/components/smooth-cursor.tsx
